+++ database/migrations/2021_09_19_154222_create_approvisionnements_table.php
@@ -18,7 +18,7 @@ class CreateApprovisionnementsTable extends Migration
             $table->integer('id_user');
             $table->string('nom');
             $table->text('description')->nullable();
-            $table->float('montant');
+            $table->double('montant');
             $table->timestamps();
         });
     }
